Home component added: hero section added
