+++ src/pages/home/components/hero/hero.css
@@ -1,0 +1,98 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@500&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
+
+#hero {
+    display: flex;
+    padding: 30px 50px;
+    background: transparent;
+}
+
+#hero div {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 30px 60px;
+    background: #F4F6F5;
+    border-radius: 23px;
+}
+
+#hero div div:first-child {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+#hero div div:nth-child(2) {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+}
+
+#hero div div:first-child .title {
+    font-weight: 900;
+    font-size: 4rem;
+    margin-left: 20px;
+    font-style: normal;
+    text-align: left;
+    text-transform: uppercase;
+    margin-top: 15.36px !important;
+    font-family: "Poppins", sans-serif;
+}
+
+#lets-bg {
+    padding-left: 20px;
+    padding-right: 50px;
+    background: white;
+}
+
+#unique-bg {
+    padding-left: 20px;
+    padding-right: 50px;
+    background: #EBD96B;
+}
+
+#hero div div:first-child .title:nth-child(2) {
+    color: black;
+    margin-top: 50px !important;
+    font-weight: 500;
+    font-size: 1.2rem;
+    font-style: normal;
+    text-align: center;
+    text-transform: none;
+    font-family: "Poppins", sans-serif;
+}
+
+#hero div div:first-child #btn-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: flex-start;
+}
+
+#hero div div:first-child #btn-container img {
+    width: 150px;
+}
+
+#hero div div:first-child #btn-container button {
+    color: white;
+    background: black;
+    border: none;
+    outline: none;
+    border-radius: 10px;
+    padding: 12px 22px;
+    font-weight: 500;
+    font-size: 0.95rem;
+    font-style: normal;
+    text-align: center;
+    text-transform: capitalize;
+    font-family: "Poppins", sans-serif;
+    margin-left: 25px;
+    cursor: pointer;
+}
+
+#hero div div:nth-child(2) img {
+    width: 100%;
+    aspect-ratio: 3/2;
+    object-fit: contain;
+}
